refactor(utils): split utils.js into several file in a folder
